pyproject.toml
[tool.poetry]
name = "nx-concorde"
version = "0.1.0"
description = "Concorde TSP solver for networkx graphs."
authors = ["Jan-Benedikt Jagusch <jan.jagusch@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
networkx = "^2.5"
numpy = "^1.19.2"
cython = "^0.29.21"
pyconcorde = { git = "https://github.com/jvkersch/pyconcorde.git" }
tsplib95 = "^0.7.1"
scipy = "^1.5.2"
pathos = "^0.2.6"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
black = "^20.8b1"
pylint = "^2.6.0"
matplotlib = "^3.3.2"
ipykernel = "^5.3.4"
profilehooks = "^1.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
